requirements.txt
pdfplumber>=0.10.0
openai>=1.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
click>=8.1.0

